requirements.txt
# 股票分析系统依赖包

# Web框架
Flask==3.0.0
Flask-CORS==4.0.0

# 数据库
duckdb==0.9.2
SQLAlchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
dbutils==3.0.0

# 数据源
akshare==1.17.1
tushare==1.3.8

# 任务调度
APScheduler==3.10.4

# 配置文件处理
PyYAML==6.0.1

# 数据处理
pandas==2.1.4
numpy==1.26.2

# 日期时间处理
python-dateutil==2.8.2

# HTTP请求
requests==2.31.0

# 工具库
python-dotenv==1.0.0
